@@ -12,15 +12,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.5",
-    "@angular/common": "^7.2.5",
-    "@angular/compiler": "^7.2.5",
-    "@angular/core": "^7.2.5",
-    "@angular/forms": "^7.2.5",
-    "@angular/http": "^7.2.5",
-    "@angular/platform-browser": "^7.2.5",
-    "@angular/platform-browser-dynamic": "^7.2.5",
-    "@angular/router": "^7.2.5",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@ng-bootstrap/ng-bootstrap": "^4.0.0",
     "@ngx-translate/core": "^11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
@@ -31,19 +30,19 @@
     "chart.js": "^2.7.3",
     "core-js": "^2.5.4",
     "moment": "^2.22.2",
-    "rxjs": "^6.4.0",
-    "zone.js": "^0.8.29"
+    "rxjs": "^6.5.5",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.6",
-    "@angular/cli": "~7.2.4",
-    "@angular/compiler-cli": "^7.2.5",
-    "@angular/language-service": "^7.2.5",
+    "@angular-devkit/build-angular": "~0.803.25",
+    "@angular/cli": "~8.3.26",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@types/chart.js": "^2.7.40",
     "@types/jasmine": "~2.8.11",
     "@types/jasminewd2": "~2.0.6",
     "@types/node": "~10.12.8",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~3.1.1",
@@ -54,6 +53,6 @@
     "protractor": "~5.4.1",
     "ts-node": "~7.0.1",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.4"
+    "typescript": "~3.5.3"
   }
 }
